Add links to RSS versions of list and category pages.
I struggled to come up with a simultaneously accessible, visible, non-ugly way to do this. Wrapping the link in a <nav> with an aria label made the best sense to me, and putting it below the sr-only <h1> and the description of the page was the best place I could finAdd links to RSS versions of list and category pages.

I struggled to come up with a simultaneously accessible, visible,
non-ugly way to do this. Wrapping the link in a <nav> with an aria label
made the best sense to me, and putting it below the sr-only <h1> and the
description of the page was the best place I could find.